removing v1.0 nodes
