+++ Java/core/src/main/java/com/amazon/randomcutforest/RCF3.java
@@ -15,22 +15,6 @@
 
 package com.amazon.randomcutforest;
 
-import static com.amazon.randomcutforest.CommonUtils.checkArgument;
-import static com.amazon.randomcutforest.CommonUtils.checkNotNull;
-import static java.lang.Math.max;
-
-import java.util.ArrayList;
-import java.util.Arrays;
-import java.util.Collections;
-import java.util.Comparator;
-import java.util.List;
-import java.util.Optional;
-import java.util.Random;
-import java.util.function.BiFunction;
-import java.util.function.BinaryOperator;
-import java.util.function.Function;
-import java.util.stream.Collector;
-
 import com.amazon.randomcutforest.anomalydetection.AnomalyAttributionVisitor;
 import com.amazon.randomcutforest.anomalydetection.AnomalyScoreVisitor;
 import com.amazon.randomcutforest.anomalydetection.DynamicAttributionVisitor;
@@ -63,10 +47,25 @@ import com.amazon.randomcutforest.store.IPointStore;
 import com.amazon.randomcutforest.store.PointStore;
 import com.amazon.randomcutforest.tree.IBoundingBoxView;
 import com.amazon.randomcutforest.tree.ITree;
-import com.amazon.randomcutforest.tree.NewRandomCutTree;
 import com.amazon.randomcutforest.tree.RandomCutTree;
 import com.amazon.randomcutforest.util.ArrayUtils;
 import com.amazon.randomcutforest.util.ShingleBuilder;
+
+import java.util.ArrayList;
+import java.util.Arrays;
+import java.util.Collections;
+import java.util.Comparator;
+import java.util.List;
+import java.util.Optional;
+import java.util.Random;
+import java.util.function.BiFunction;
+import java.util.function.BinaryOperator;
+import java.util.function.Function;
+import java.util.stream.Collector;
+
+import static com.amazon.randomcutforest.CommonUtils.checkArgument;
+import static com.amazon.randomcutforest.CommonUtils.checkNotNull;
+import static java.lang.Math.max;
 
 /**
  * The RandomCutForest class is the interface to the algorithms in this package,
@@ -303,8 +302,8 @@ public class RCF3 {
         IStateCoordinator<Integer, float[]> stateCoordinator = new PointStoreCoordinator<>(tempStore);
         ComponentList<Integer, float[]> components = new ComponentList<>(numberOfTrees);
         for (int i = 0; i < numberOfTrees; i++) {
-            ITree<Integer, float[]> tree = new NewRandomCutTree.Builder<>().capacity(sampleSize)
-                    .outputAfter(outputAfter).randomSeed(random.nextLong()).pointStoreView(tempStore)
+            ITree<Integer, float[]> tree = new RandomCutTree.Builder<>().capacity(sampleSize).outputAfter(outputAfter)
+                    .randomSeed(random.nextLong()).pointStoreView(tempStore)
                     .boundingBoxCacheFraction(boundingBoxCacheFraction).build();
 
             IStreamSampler<Integer> sampler = CompactSampler.builder().capacity(sampleSize).timeDecay(timeDecay)
