+++ Java/core/src/main/java/com/amazon/randomcutforest/store/NodeStore.java
@@ -36,8 +36,6 @@ import java.util.Arrays;
  * is a Node), then this class is analogous to a column-oriented database of
  * Nodes.
  *
- * Note that a NodeStore does not store instances of the
- * {@link com.amazon.randomcutforest.tree.Node} class.
  */
 public class NodeStore implements INodeStore {
