+++ Java/core/src/main/java/com/amazon/randomcutforest/store/SmallNodeStore.java
@@ -37,8 +37,6 @@ import java.util.Arrays;
  * is a Node), then this class is analogous to a column-oriented database of
  * Nodes.
  *
- * Note that a NodeStore does not store instances of the
- * {@link com.amazon.randomcutforest.tree.Node} class.
  */
 public class SmallNodeStore implements INodeStore {
