+++ Java/core/src/main/java/com/amazon/randomcutforest/tree/HyperTree.java
@@ -16,9 +16,10 @@
 package com.amazon.randomcutforest.tree;
 
 import static com.amazon.randomcutforest.CommonUtils.checkArgument;
-import static com.amazon.randomcutforest.tree.Cut.isLeftOf;
+import static com.amazon.randomcutforest.tree.AbstractNodeStore.Null;
 
 import java.util.ArrayList;
+import java.util.Arrays;
 import java.util.List;
 import java.util.Random;
 import java.util.function.Function;
@@ -40,31 +41,41 @@ public class HyperTree extends RandomCutTree {
         this.gVecBuild = builder.gVec;
     }
 
-    public void makeTree(List<double[]> list, int seed) {
+    public void makeTree(List<Integer> list, int seed) {
         // this function allows a public call, which may be useful someday
-        if (list.size() > 0) {
-            // dimensions = list.get(0).length;
-            root = makeTreeInt(list, seed, 0, this.gVecBuild);
+        if (list.size() > 0 && list.size() < numberOfLeaves + 1) {
+            int[] leftIndex = new int[numberOfLeaves - 1];
+            int[] rightIndex = new int[numberOfLeaves - 1];
+            Arrays.fill(leftIndex, numberOfLeaves - 1);
+            Arrays.fill(rightIndex, numberOfLeaves - 1);
+            int[] cutDimension = new int[numberOfLeaves - 1];
+            float[] cutValue = new float[numberOfLeaves - 1];
+            root = makeTreeInt(list, seed, 0, this.gVecBuild, leftIndex, rightIndex, cutDimension, cutValue);
+            nodeStore = AbstractNodeStore.builder().storeSequencesEnabled(false).pointStoreView(pointStoreView)
+                    .dimensions(dimension).capacity(numberOfLeaves - 1).leftIndex(leftIndex).rightIndex(rightIndex)
+                    .cutDimension(cutDimension).cutValues(cutValue).build();
+            // the cuts are specififed; now build tree
+            for (int i = 0; i < list.size(); i++) {
+                addPoint(list.get(i), 0L);
+            }
         } else {
-            root = null;
+            root = Null;
         }
     }
 
-    private Node makeTreeInt(List<double[]> pointList, int seed, int level,
-            Function<IBoundingBoxView, double[]> vecBuild) {
+    private int makeTreeInt(List<Integer> pointList, int seed, int firstFree,
+            Function<IBoundingBoxView, double[]> vecBuild, int[] left, int[] right, int[] cutDimension,
+            float[] cutValue) {
 
         if (pointList.size() == 0)
-            return null;
+            return Null;
 
-        BoundingBox thisBox = new BoundingBox(pointList.get(0));
+        BoundingBoxFloat thisBox = new BoundingBoxFloat(pointStoreView.get(pointList.get(0)));
         for (int i = 1; i < pointList.size(); i++) {
-            thisBox = thisBox.getMergedBox(pointList.get(i));
+            thisBox = (BoundingBoxFloat) thisBox.getMergedBox(pointStoreView.get(pointList.get(i)));
         }
         if (thisBox.getRangeSum() <= 0) {
-            double[] first = pointList.get(0);
-            Node x = new Node(first);
-            x.setMass(pointList.size());
-            return x;
+            return pointList.get(0) + nodeStore.getCapacity() + 1;
         }
 
         Random ring = new Random(seed);
@@ -72,23 +83,24 @@ public class HyperTree extends RandomCutTree {
         int rightSeed = ring.nextInt();
         Cut cut = getCut(thisBox, ring, vecBuild);
 
-        List<double[]> leftList = new ArrayList<>();
-        List<double[]> rightList = new ArrayList<>();
+        List<Integer> leftList = new ArrayList<>();
+        List<Integer> rightList = new ArrayList<>();
 
         for (int j = 0; j < pointList.size(); j++) {
-            if (isLeftOf(pointList.get(j), cut)) {
+            if (nodeStore.leftOf((float) cut.getValue(), cut.getDimension(), pointStoreView.get(pointList.get(j)))) {
                 leftList.add(pointList.get(j));
             } else
                 rightList.add(pointList.get(j));
 
         }
-        Node leftNode = makeTreeInt(leftList, leftSeed, level + 1, vecBuild);
-        Node rightNode = makeTreeInt(rightList, rightSeed, level + 1, vecBuild);
-        Node thisNode = new Node(leftNode, rightNode, cut, thisBox);
-        leftNode.setParent(thisNode);
-        rightNode.setParent(thisNode);
-        thisNode.setMass(pointList.size());
-        return thisNode;
+        int leftIndex = makeTreeInt(leftList, leftSeed, firstFree + 1, vecBuild, left, right, cutDimension, cutValue);
+        int rightIndex = makeTreeInt(rightList, rightSeed, firstFree + leftList.size(), vecBuild, left, right,
+                cutDimension, cutValue);
+        left[firstFree] = Math.min(leftIndex, numberOfLeaves - 1);
+        right[firstFree] = Math.min(rightIndex, numberOfLeaves - 1);
+        cutDimension[firstFree] = cut.getDimension();
+        cutValue[firstFree] = (float) cut.getValue();
+        return firstFree;
     }
 
     private Cut getCut(IBoundingBoxView bb, Random ring, Function<IBoundingBoxView, double[]> vecSeparation) {
@@ -99,31 +111,28 @@ public class HyperTree extends RandomCutTree {
         double rangeSum = 0;
         double[] vector = vecSeparation.apply(bb);
         for (int i = 0; i < bb.getDimensions(); i++) {
+            vector[i] = (float) vector[i];
             rangeSum += vector[i];
         }
 
         double breakPoint = dimf * rangeSum;
+        float cutValue = 0;
         for (int i = 0; i < bb.getDimensions(); i++) {
             double range = vector[i];
             if (range > 0) {
                 if ((breakPoint > 0) && (breakPoint <= range)) {
                     td = i;
+                    cutValue = (float) (bb.getMinValue(td) + bb.getRange(td) * cutf);
+                    if (cutValue == bb.getMaxValue(td)) {
+                        cutValue = (float) bb.getMinValue(td);
+                    }
                 }
                 breakPoint -= range;
             }
         }
 
         checkArgument(td != -1, "Pivot selection failed.");
-        return new Cut(td, bb.getMinValue(td) + bb.getRange(td) * cutf);
-    }
-
-    public double[] addPoint(double[] point, long sequenceIndex) {
-        return point;
-    }
-
-    @Override
-    public double[] deletePoint(double[] point, long sequenceIndex) {
-        return point;
+        return new Cut(td, cutValue);
     }
 
     public static class Builder extends RandomCutTree.Builder<Builder> {
