+++ Java/core/src/main/java/com/amazon/randomcutforest/tree/NodeStoreLarge.java
@@ -35,8 +35,6 @@ import com.amazon.randomcutforest.store.IndexIntervalManager;
  * is a Node), then this class is analogous to a column-oriented database of
  * Nodes.
  *
- * Note that a NodeStore does not store instances of the
- * {@link com.amazon.randomcutforest.tree.Node} class.
  */
 public class NodeStoreLarge extends AbstractNodeStore {
